Fix to flipeditor import; removed more md-[icon] prefixes
