+++ .gitignore
@@ -7,3 +7,4 @@
 /src/utils/translations/current.json
 /web-build/*
 /dist/*
+.env*.local
