+++ package.json
@@ -91,7 +91,7 @@
     "react-dom": "18.2.0",
     "react-native": "0.73.6",
     "react-native-circular-progress": "1.3.7",
-    "react-native-flip-editor": "github:educational-resources-and-services/react-native-flip-editor#68770463c4159267b5041a081367940adb9f05ab",
+    "react-native-flip-editor": "github:educational-resources-and-services/react-native-flip-editor#7bb67786ef504ca5694b757f915b0e883e6d7f26",
     "react-native-gesture-handler": "~2.14.0",
     "react-native-reanimated": "~3.6.2",
     "react-native-safe-area-context": "4.8.2",
@@ -113,7 +113,7 @@
     "sentry-expo": "~7.2.0",
     "stream": "0.0.2",
     "timers": "0.1.1",
-    "uuid": "3.4.0",
+    "uuid": "^9.0.1",
     "victory": "34.1.3",
     "victory-native": "34.1.0",
     "vm-browserify": "^1.1.2",
