+++ src/components/major/AudioPlayer.js
@@ -205,8 +205,8 @@ const AudioPlayer = ({
 
   const BackFiveIcon = useCallback(({ style }) => <Icon name='replay-5' pack="material" style={styles.icon} />, [])
   const ForwardFiveIcon = useCallback(({ style }) => <Icon name='forward-5' pack="material" style={styles.icon} />, [])
-  const PlayIcon = useCallback(({ style }) => <Icon name='md-play' style={styles.bigIcon} />, [])
-  const PauseIcon = useCallback(({ style }) => <Icon name='md-pause' style={styles.bigIcon} />, [])
+  const PlayIcon = useCallback(({ style }) => <Icon name='play' style={styles.bigIcon} />, [])
+  const PauseIcon = useCallback(({ style }) => <Icon name='pause' style={styles.bigIcon} />, [])
 
   return (
     <View style={styles.container}>
